Re #33 Break down the level view template
The template is broken down into what runs only as part of the registration process and what is used to display the user fields.
The common part is now its own template file (default_fields).
The signup.js file now doesn't expect the coupon, payment method and price fields to always be part of the form