use valid pre-FOF3 functions & fix the duplicates
use valid pre-FOF3 functions rather than directly copying the dev branch
fix the duplicates that happen on new subs with slaves when run integrations is clicked (i.e. onAKUserRefresh)

TODO
make sure when syncing the parent information to the slave that the state and payment status is copied. otherwise the slave may never be vaild (since it wont be active)