Add clone before F0FModel::getTmpInstance
As per F0F documentation, "// If you forget the clone you will spend some not-so-fun hours debugging erratic Behavior" (Yep, that happened)
This PR fixes the issue with lines 433 and 434 where the variables $from and $to were resulting in the same record being passed to copySubscriptionInformation. With this PR slave subs are finally valid and active in AS 4. For good measure, following the F0F documentation all instances of F0FModel::getTmpInstance were changed to "clone F0FModel::getTmpInstance()" 
https://www.akeebabackup.com/documentation/fof/foftable-simpleorm.html