Use `!empty()` rather than `isset()`
We are working with DataModel Objects and DataModel properties and not just variables. If I understand it right; when isset() is used on a DataModel property it detects if the property has been initialized, and will return false since all DataModel properties are always considered set even when empty. 

In short, don't use `isset()` use `!empty()`